Interpolate enhancements
Allow for a guess to be specified. Currently set to 0.5 as subdevider. The
commented out version has issues and i haven't found yet why it doesn't work.

Also add a check to terminate directly if one of the two begin points is
already the desired distance to the zero, stop directly.