Re-implement the SCAD output module.
Note that there's a really nice pretty printer inside that module
just waiting for the Reader's context to change from ℝ to (ℝ,Int) and a
few changes to call.

Otherwise, this is the same story as all of the other rewrites - no
more string concatenation and slightly less awkward parameter passing.

Note that this does introduce a dependency on mtl, which may not be
popular. This would also work with transformers, so...

(Commit 6/6 of my quest to add efficient GHCLive support for ImplicitCAD.)