Be more explicit about endianness for stl export.
Note that I'm explicitly using Float for the signature of float32LE
instead of ℝ - if ℝ should change to Double in the future, we'll type
error, not produce bad stl. storable-endian doesn't seem to add any
significant performance issues - although it (the package) does use
a depreciated function (unsafeIO from the wrong module).