Link extopenscad against the library
Previously, cabal built the library modules twice, once for the library
itself, and once for extopenscad. Now the library is built once, and
extopenscad is linked against it.

This required moving extopenscad to its own subdirectory, and a small
change in its imports to only make use of exposed modules.