Add basic unit tests and benchmarks for parser
The tests uncover a problem in the parser handling the ternary operator,
the testcases that would fail are marked as pending (see 'ternaryIssue'
in the code).

The benchmarks indicate that the parser is unreasonably inefficient;
here are is the result summary for my machine (Intel i5-6600K, 3.5GHz):

- 5000 line comments, 80 chars per line: ~50ms
- 500 block comments spanning 10 lines, 40 chars per line: ~58ms
- 100 assignments ("x = (foo + bar);"): ~268ms
- List of 1000 integers: ~1.1s
- Arithmetic expressions with nesting level 3: ~2.3s

Especially the integer list and arithmetic benchmark times seems way
worse than they could/should be.

Note that I already have a refactored parser in preparation that will
hopefully do significantly better.