Change the type of differenceR to include a mandatory base shape
The first element of the list argument to `differenceR` is special, and
acts as the shape from which the others are subtracted. The previous
behavior of exporting `differenceR r []` (or even `differenceR [x]`) was
to crash due to `head` and `tail` being partial functions.

From this analysis, it's clear that the "shape to remove from" is
a *mandatory* argument to `differenceR`, and as such should be
explicitly moved out of the list args.