Add golden test machinery
This PR adds a new GoldenTestSpec, which exposes a `golden` combinator
for building golden tests. These tests take a SymbolicObj3, and will
raytrace it, checking the resulting STL against a known-good version.
The result is a robust system for ensuring mesh generation doesn't
subtly change when working on the underlying infrastructure.