Splitting up StateE and parameterising ImplicitCadM
StateE has been split up along the same lines as StateC.
ImplicitCadM has been fully parameterised so it can be used for StateE,
allowing us to have a single transformer core stack for all of the
reader/writer/state handling, dropping in types as needed.
Changing most references to ImplicitCadM to StateC, as it is a type
alias that has the expected types parameterised.