Adding basic support for `rands`
Adding basic support for the rands function, without support for
seeding the rang explicitly. It is using the global prng.

This also adds the `OIO` OVal type, supporting IO actions. This is
nessacary for using the global rng.

Adding support for indexing into OIO values.

Adding some expression tests for `rands`, checking that values are in
the desired range and count.