Fixing an off by one error in Expr variable evaluation.
Reversing the order of pattern names and values when evaluating Exprs,
and changing how variable lookup is performed to ensure that if a
variable is used it always has a value associated with it.

Adding tests to check that let bindings work as expected when nested and
overloading names in scope.