Improving `shell`
Fixing an issue in `shell` where it was causing the object to increase
in size, proportional to the width value passed into shell.

`shell` now works by taking the difference of the original object and
the same object with a negative outset applied. This process does not
change the outside dimentions of the resulting object when compared to
the original.

Adding a new golden test that can also be used to visually check that
shell isn't changing outside dimentions.
Updating the wheel-well STL file as it depends on `shell`