Issue #450. Making torus and ellipsoid primitive objects.
Switching the definitions of torus and ellipsoid to be SymbolicObj3
primitives rather than being `implict`ly defined. The maths functions
haven't changed, only been moved around.

One test was updated as some of the STL values changed, but to the eye
it looks similar, and a random check of values indicated that they were
small.