Fixing relative imports for extopenscad.
When exteopnscad is called with a source filepath outside of the current
directory, relative filepaths such as "../config.scad" would fail to
read if no such file relative to the _programs_ directory, not the
source file, existed.

This change uses the directory of the initial source file as the
location for relative file lookups.